requirements.txt
bleach==3.1.0
certifi==2016.2.28
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.8
django-ckeditor==5.6.1
django-crequest==2018.5.11
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-fancybox==0.1.4
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-mdeditor==0.1.13
django-model-utils==3.1.2
django-ranged-response==0.2.0
django-reversion==3.0.3
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.1
jdcal==1.4
jieba==0.39
Markdown==3.0.1
odfpy==1.4.0
openpyxl==2.6.0
Pillow==5.4.1
Pygments==2.3.1
PyMySQL==0.9.3
python-memcached==1.59
pytz==2017.2
PyYAML==3.13
six==1.12.0
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
webencodings==0.5.1
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
